% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_FDR.R
\name{add_fdr}
\alias{add_fdr}
\title{add_fdr}
\usage{
add_fdr(df, snpcol = "PP4")
}
\arguments{
\item{column}{name of posterior probability to use for FDR}
}
\value{
data frame with added column
}
\description{
FDR using PPA
}
\details{
Add a column with FDR of colocalization results
}
\author{
Claudia Giambartolomei
}
\keyword{internal}

