% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moloc.R
\name{adjust_bfs}
\alias{adjust_bfs}
\title{adjust_bfs}
\usage{
adjust_bfs(listData, prior_var = "default")
}
\arguments{
\item{listData}{A list of data frames to be analyzed. Each data frame
must contain columns named SNP (the SNP name consistent across all the data frames), 
BETA, 
SE}

\item{prior_var}{One or more numbers for the prior variance of the ABF.}

\item{compute_sdY}{Estimate standard deviation of Y. If false assume sdY is 1}
}
\value{
An array containing the log adjusted Bayes Factors for each SNP and
        for each SNP and each configuration combination
}
\description{
Adjusted Bayes factors for each SNP and each single association and combinations of
sharing/not sharing of causal variants across the datasets in the input data
}
\examples{
ABF <- adjust_bfs(listData, prior_var=c(0.01, 0.1, 0.5))
to view all configuration combinations for a SNP called "bp38878088"
ABF["bp38878088",,]

}
\author{
Claudia Giambartolomei, Jimmy Liu
}
\keyword{internal}

