% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moloc_overlap.R
\name{adjust_bfs_overlap}
\alias{adjust_bfs_overlap}
\title{adjust_bfs}
\usage{
adjust_bfs_overlap(listData, prior_var = 0.15^2, overlap = FALSE,
  compute_sdY = FALSE, from_p = FALSE)
}
\arguments{
\item{listData}{A list of data frames to be analyzed. Each data frame
must contain columns named SNP (the SNP name consistent across all the data frames), 
BETA, 
SE}

\item{prior_var}{One or more numbers for the prior variance of the ABF.}

\item{overlap}{Logical, do the individuals in the datasets overlap?}

\item{compute_sdY}{Estimate standard deviation of Y. If false assume sdY is 1}
}
\value{
An array containing the log adjusted Bayes Factors for each SNP and
        for each SNP and each configuration combination
}
\description{
Adjusted Bayes factors for each SNP and each single association and combinations of
sharing/not sharing of causal variants across the datasets in the input data
}
\examples{
ABF <- adjust_bfs_overlap(listData, overlap=FALSE, prior_var=c(0.01, 0.1, 0.5))
ABF <- adjust_bfs(listData=listData, overlap=overlap, prior_var=prior_var, compute_sdY=compute_sdY, from_p=from_p)
to view all configuration combinations for a SNP called "bp38878088"
ABF["bp38878088",,]

}
\author{
Jimmy Liu, Claudia Giambartolomei
}
\keyword{internal}

