% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_alleles_betas.R, R/functions_moloc.R
\name{change_indels}
\alias{change_indels}
\title{change_indels}
\usage{
change_indels(data)

change_indels(data)
}
\arguments{
\item{data, }{a data frame with columns A1 and A2}

\item{data, }{a data frame with columns A1 and A2}
}
\value{
data frame

data frame
}
\description{
Change indel format to I/D

Change indel format to I/D
}
\author{
Claudia Giambartolomei

Claudia Giambartolomei
}
\seealso{
Other align.functions: \code{\link{complement_snp}},
  \code{\link{match_alleles}}

Other align.functions: \code{\link{complement_snp}},
  \code{\link{match_alleles}}
}
\keyword{internal}

