% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_moloc.R
\name{coloc.genome}
\alias{coloc.genome}
\title{coloc.genome}
\usage{
coloc.genome(listData, bed, prefix = "pref", save.SNP.info = FALSE,
  cores = 20, have_alleles = TRUE, bychrpos = TRUE,
  prior_var = "default", priors = c(1e-04, 1e-06, 1e-07), min_nsnps = 50,
  takelog = FALSE, write = TRUE, outfolder = "test", forcePVAL = FALSE)
}
\arguments{
\item{listData, }{a list of data frames in this order: gwas, and other QTLs
The data frames need to have columns: "SNP" or "CHR", "POS"; "BETA", "SE"; "N", "MAF" (to estimate sdY)
                                          if a case control: "Ncases"
                                       If eqtl/mqtl: "ProbeID"
                                       Optionally, "A1", "A2" if want to match alleles}

\item{bed, }{a data frame with "CHR", "START", "STOP", and 
"ProbeID" matching either eqtl or mqtl}

\item{prefix, }{character, name to give to the output file}

\item{save.SNP.info, }{logical, should info of each SNP be saved? 
If this option is true, beware that it will takes up a lot of time and space.}

\item{cores, }{number. See foreach package.}

\item{have_alleles, }{logical. If TRUE, matches alleles to be the same as gwas data.}

\item{bychrpos, }{logical. If TRUE, uses "CHR", "POS" columns 
as SNP id to match across data frames;
If TRUE, make sure not to have another column called "SNP".}

\item{prior_var, }{numeric vector specifying any number of values;
These will be used to average the ABF computation across different prior variances.}

\item{priors, }{numeric vectors with three numbers; 
First number is the prior of associaion of a SNP with any of the traits;
Second number is the prior of a SNP being associated with two traits;
Third number is the prior of a SNP being associated with three traits.}

\item{min_nsnps, }{number. The minimum number of matching SNPs across 
the three data frames to consider.}

\item{outfolder, }{path of folder where the results will be written to}
}
\value{
results in outfolder: text file with colocalization results for each locus; 
        text file with removed snps; 
        if save.SNP.info = TRUE, a folder called "coloc.output.perSNP"
        will save a file per locus with SNP information such as ABF (one row per matching SNP)
}
\description{
Genome-wide co-localization with three traits
}
\examples{
library(mvtnorm)
library(foreach)
library(doParallel)
library(data.table)

Create bed file with combination of ProbeIDs
library(GenomicRanges)
DT <- as.data.table(data_genome[[3]])
methyl_table <- DT[, list(CHR= unique(CHR), START = min(POS), STOP = max(POS)), by = ProbeID]
bed1.gr <- GRanges(seqnames = bed$CHR,IRanges(start = bed$START, end= bed$STOP))
bed2.gr <- GRanges(seqnames = methyl_table$CHR,IRanges(start =methyl_table$START, end= methyl_table$STOP))
my.overlap <- findOverlaps(query = bed2.gr, subject = bed1.gr)
bed = cbind(bed[my.overlap@subjectHits,], methyl_table[my.overlap@queryHits,])
bed = bed[,c(1,6,7,8,9,10)]
Setting the prior_var to "default" uses coloc defaults
coloc.genome(listData = data_genome, bed, cores=1, have_alleles=TRUE, bychrpos=TRUE, prior_var=NULL, priors=c(1e-04, 1e-06, 1e-07), min_nsnps = 50, write=FALSE, outfolder = "test")

}
\author{
Claudia Giambartolomei
}

