% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_alleles_betas.R, R/functions_moloc.R
\name{complement_snp}
\alias{complement_snp}
\title{complement_snp}
\usage{
complement_snp(x)

complement_snp(x)
}
\arguments{
\item{x}{character}

\item{x}{character}
}
\value{
character, complement

character, complement
}
\description{
Find complement SNP

Find complement SNP
}
\examples{
complement_snp("A")

complement_snp("A")

}
\author{
James Boocock

James Boocock
}
\seealso{
Other align.functions: \code{\link{change_indels}},
  \code{\link{match_alleles}}

Other align.functions: \code{\link{change_indels}},
  \code{\link{match_alleles}}
}
\keyword{internal}

