% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moloc.R
\name{config_coloc}
\alias{config_coloc}
\title{config_coloc}
\usage{
config_coloc(ABF, n_files, priors)
}
\arguments{
\item{ABF}{is an array containing the single and coloc logBFs}

\item{n_files}{is one number, i.e. the number of traits to be analyzed}

\item{priors}{are numbers, the prior for one variant to be associated with 1 trait and with each additional trait}
}
\value{
A data frame containing the likelihoods and posteriors for each configuration combination
}
\description{
Likelihood frame and posterior probability for combinations of sharing/not sharing
of causal variants across the datasets in the input data
}
\examples{
lkl <- config_coloc(ABF, n_files=3, priors=c(1e-04, 1e-06, 1e-07))

}
\author{
Jimmy Liu, Claudia Giambartolomei
}
\keyword{internal}

