% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_genome.R
\docType{data}
\name{data_genome}
\alias{data_genome}
\title{Example for genome wide analysis of three traits}
\format{A list of three data frames with summary statistics for each data: 
the biomarker data frame with 746 rows and 11 columns;
the eqtl data frame with 397 rows and 11 columns; 
the methylation data frame with 12951 and 10 columns.
\describe{
  \item{SNP}{SNP names, common to all data frames}
  \item{BETA}{Effect size estimate from regression, if OR must give the logOR}
  \item{SE}{Standard Error}
  \item{z}{Z-score}
  \item{PVAL}{P-value}
  \item{N}{Sample size}
  \item{MAF}{Minor allele frequency}
  ...
}}
\usage{
data_genome
}
\description{
A list of data frames containing summary statistics from three indendent datasets,
for example GWAS, expression QTL, methylation QTL
}
\keyword{datasets}

