% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_single.R
\docType{data}
\name{data_single}
\alias{data_single}
\title{List including three data frames with summary statistics for three traits}
\format{A list of three data frames, each with 831 rows and 7 variables:
\describe{
  \item{SNP}{SNP names, common to all data frames}
  \item{BETA}{Effect size estimate from regression, if OR must give the logOR}
  \item{SE}{Standard Error}
  \item{z}{Z-score}
  \item{PVAL}{P-value}
  \item{N}{Sample size}
  \item{MAF}{Minor allele frequency}
  ...
}}
\usage{
data_single
}
\description{
A list of data frames containing summary statistics from three indendent datasets,
for example GWAS, expression QTL, methylation QTL
}
\keyword{datasets}

