% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmvnorm.R
\name{dmvnorm}
\alias{dmvnorm}
\title{dmvnorm}
\usage{
dmvnorm(x, mean = rep(0, p), sigma = diag(p), log = FALSE)
}
\arguments{
\item{x, }{numeric}

\item{mean, }{numeric}

\item{sigma, }{numeric}

\item{log, }{logical}
}
\value{
matrix
}
\description{
dmvnorm from dmvnorm package
}
\details{
Internal function
}
\examples{
dmvnorm(betas, means, sigma_h1)
}
\author{
dmvnorm package
}
\keyword{internal}

