% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_Homo_sapiens.GRCh37.70.processed.R
\docType{data}
\name{ensemble_Homo_sapiens.GRCh37.70.processed}
\alias{ensemble_Homo_sapiens.GRCh37.70.processed}
\title{Bed file with Ensembl genes, start position; stop position; for genome-wide analysis.
If using this file, it is recemmended to add +/- 50000 for methylation analysis, or 200000 for eQTL/GWAS analysis.}
\format{A data frame with 3 variables:
\describe{
  \item{ProbeID}{Name of Genes or methylation Probes, as listed in the eqtl/mqtl summary stats}
  \item{CHR}{CHR}
  \item{START}{START of region}
  \item{STOP}{STOP of region}
  ...
}}
\usage{
ensemble_Homo_sapiens.GRCh37.70.processed
}
\description{
A tab delimited text file with ProbeID and regions to consider
for the moloc analysis
}
\keyword{datasets}

