% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_causalM}
\alias{get_causalM}
\title{get_combos}
\usage{
get_causalM(mm, v)
}
\arguments{
\item{x}{matrix}
}
\value{
matrix with causal configurations
}
\description{
Internal function, get_causalM
}
\details{
Get causal matrix from binary matrix; v is the probability per SNP
}
\author{
Claudia Giambartolomei
}
\keyword{internal}

