% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get_priors}
\alias{get_priors}
\title{get_priors}
\usage{
get_priors(v, n_files)
}
\arguments{
\item{v, }{numeric vector of probabilities for each SNP included in the analysis}

\item{n_files, }{number of traits analyzed, for now can be either 2 or 3}

\item{priors, }{numeric vectors with three numbers;}
}
\value{
probabilities for each hypotheses in moloc (5 for 2 traits; 15 for three traits).
}
\description{
priors from causal probabilities
}
\examples{
v = runif(1:1000)
n_files = 3
get_priors(v, n_files)

}
\author{
Claudia Giambartolomei
}

