% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moloc_overlap.R
\name{get_psd}
\alias{get_psd}
\title{get_psd}
\usage{
get_psd(A)
}
\arguments{
\item{A}{matrix}
}
\description{
Internal function, get_psd
}
\details{
This function attempts to make a matrix positive definite
}
\author{
Claudia Giambartolomei, Jimmy Liu
}
\keyword{internal}

