% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moloc_overlap.R
\name{get_stats}
\alias{get_stats}
\title{stats_p_cc}
\usage{
get_stats(x)
}
\arguments{
\item{s}{proportion of samples that are cases}

\item{f}{minor allele freq}

\item{N}{sample number}
}
\value{
variance of MLE beta
}
\description{
Internal function, stats_p_cc
}
\details{
Internal function
}
\keyword{internal}

