% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moloc_overlap.R
\name{is_pos_def}
\alias{is_pos_def}
\title{is_pos_def}
\usage{
is_pos_def(A)
}
\arguments{
\item{A}{matrix}
}
\description{
Internal function, is_pos_def
}
\details{
This function tests if a matrix is positive definite
}
\author{
Claudia Giambartolomei, Jimmy Liu
}
\keyword{internal}

