% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moloc_overlap.R
\name{make_sigma}
\alias{make_sigma}
\title{make_sigma}
\usage{
make_sigma(cor_mat, v, w)
}
\arguments{
\item{cor_mat}{matrix, correlation between the traits for each SNP
If assuming no overlap, cor_mat is identity matrix.}
}
\value{
matrix of ABF adjusted for correlations
}
\description{
Internal function, make_sigma
}
\details{
This function creates Wakefield's ABF for each SNP, taking account of the correlation between traits
}
\author{
Jimmy Liu
}
\keyword{internal}

