% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_alleles_betas.R, R/functions_moloc.R
\name{match_alleles}
\alias{match_alleles}
\title{match_alleles}
\usage{
match_alleles(data, A1.ref = "A1.ref", A2.ref = "A2.ref", A1.data = "A1",
  A2.data = "A2", BETA.data = "BETA", flip = TRUE)

match_alleles(data, A1.ref = "A1.ref", A2.ref = "A2.ref", A1.data = "A1",
  A2.data = "A2", BETA.data = "BETA", flip = TRUE)
}
\arguments{
\item{data, }{is a data.frame with columns A1 and A2, and reference alleles 
that need to match}

\item{A1.ref}{column names of ref alleles for A1 in the same data frame}

\item{A2.ref}{column names of ref alleles for A2 in the same data frame}

\item{flip}{logical, whether to flip the betas if the alleles are flipped}

\item{data, }{is a data.frame with columns A1 and A2, and reference alleles 
that need to match}

\item{A1.ref}{column names of ref alleles for A1 in the same data frame}

\item{A2.ref}{column names of ref alleles for A2 in the same data frame}

\item{flip}{logical, whether to flip the betas if the alleles are flipped}
}
\value{
data frame

data frame
}
\description{
Match alleles to a reference and flip betas

Match alleles to a reference and flip betas
}
\author{
Claudia Giambartolomei, James Boocock

Claudia Giambartolomei, James Boocock
}
\seealso{
Other align.functions: \code{\link{change_indels}},
  \code{\link{complement_snp}}

Other align.functions: \code{\link{change_indels}},
  \code{\link{complement_snp}}
}
\keyword{internal}

