% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moloc-package.R
\docType{package}
\name{moloc-package}
\alias{moloc-package}
\title{Colocalisation tests of two genetic traits}
\description{
Performs the colocalisation tests of multiple traits (manuscript in preparation)
}
\details{
\code{moloc.test()} tests for colocalisation and returns an object of class
\code{moloc}.
}
\author{
Claudia Giambartolomei <claudia.giambartolomei@gmail.com>
}
\references{
Giambartolomei et al (2014).  Bayesian Test for Colocalisation between Pairs of Genetic Association Studies Using Summary Statistics.  PLOS Genet e1004383.

\url{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC4022491/}
}

