% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{nested_lapply_args}
\alias{nested_lapply_args}
\title{nested_lapply_args}
\usage{
nested_lapply_args(data, fun, v)
}
\description{
Internal function, nested_lapply_args
}
\author{
Claudia Giambartolomei
}
\keyword{internal}

