% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moloc.R
\name{snp_ppa}
\alias{snp_ppa}
\title{snp_ppa}
\usage{
snp_ppa(ABF, n_files, config_ppas)
}
\arguments{
\item{ABF}{is an array containing the single and coloc logBFs}

\item{n_files}{is one number, i.e. the number of traits to be analyzed}

\item{priors}{is the prior for one variant to be associated with 1 trait and with each additional trait}
}
\value{
A data frame containing the likelihoods and posteriors for each configuration combination
}
\description{
Posterior probability that each SNP is THE causal variant for a shared signal
}
\examples{
snp <- snp_ppa(ABF, n_files=3, config_ppas=lkl[[1]])

}
\author{
Jimmy Liu, Claudia Giambartolomei
}
\keyword{internal}

