% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moloc_overlap.R
\name{stats_p}
\alias{stats_p}
\title{Internal function, stats_p}
\usage{
stats_p(p, n, maf)
}
\arguments{
\item{p}{p value}

\item{f}{MAF}

\item{type}{"quant" or "cc"}

\item{N}{sample size}

\item{s}{proportion of samples that are cases, ignored if type=="quant"}

\item{suffix}{suffix to append to column names of returned data.frame}
}
\value{
data.frame containing beta, SE
}
\description{
Internal function, stats_p
}
\details{
Get beta, SE from p
}
\author{
Claudia Giambartolomei
}
\keyword{internal}

