% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moloc_overlap.R
\name{stats_p_cc}
\alias{stats_p_cc}
\title{stats_p_cc}
\usage{
stats_p_cc(p, n, maf, s)
}
\arguments{
\item{s}{proportion of samples that are cases}

\item{f}{minor allele freq}

\item{N}{sample number}
}
\value{
variance of MLE beta
}
\description{
Internal function, stats_p_cc
}
\details{
Internal function
}
\examples{
stats = stats_p_cc(listData[[1]]$PVAL, listData[[1]]$N, listData[[1]]$F, listData[[1]]$Ncases/listData[[1]]$N)
}
\author{
Claudia Giambartolomei
}
\keyword{internal}

